Breaking mphys script into multiple
